Clearly seperate ARGV parsing, Compile and Execution phases.
Handling ARGV via Getopts during Require is not very nice, and basically
prohibits this module being used anywhere, even in tests, because it
will expect arguments to be passed, even if you're not calling run.

So this breaks things into more logical stages.

During loading/compile, nothing happens.

->new_with_opts then handles @ARGV via Getopt etc, creating a
configuration hash.

This then calls ->new with the determined configuration to create an
object internally.

Then ->run is called on this object, instead of as a class method,
and subsequent data access is via this configuration instead of via
package lexicals.